用户注册认证流程

+++ src/main/java/com/miaosha/service/model/UserModel.java
@@ -5,7 +5,7 @@ import com.miaosha.controller.viewobject.UserVO;
 public class UserModel {
     private Integer id;
     private String name;
-    private Integer gender;
+    private Byte gender;
     private Integer age;
     private String telphone;
 
@@ -35,11 +35,11 @@ public class UserModel {
         this.name = name;
     }
 
-    public Integer getGender() {
+    public Byte getGender() {
         return gender;
     }
 
-    public void setGender(Integer gender) {
+    public void setGender(Byte gender) {
         this.gender = gender;
     }
